Moved mandatoryDirectivesForFields
